src/main/java/com/poohdev/booksharing/controller/PropertyController.java
package com.poohdev.booksharing.controller;

import com.poohdev.booksharing.service.PropertyService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/property")
public class PropertyController {

    private final PropertyService propertyService;

    public PropertyController(PropertyService propertyService) {
        this.propertyService = propertyService;
    }

    @GetMapping("/app-name")
    public String getApplicationName() {
        return propertyService.getApplicationName();
    }
}
